Major forked thold merge to Cacti core thold plugin
For those of you watching, you may have thought I was never going to commit again.  Well, you were wrong :+1:  I'm alive and still working on Cacti and a number of plugins.  Just not in a very public way.  You may be aware of my other work, and if so, that's great!

My current employer is allowing me to port a number of thold changes back to the Cacti community from our forked version.
 It was a necessity since the Cacti core version of the plugin has advanced so much since our forked version was first created.

Of the commits that will be forthcoming, this is the first.  In this first commit, I'm adding the following:

* Adding some additional hooks to the Notification List functionality for one of my non-public plubins.
* Adding Acknowledgment functionality to thold
* Allowing optional per threshold and per template specific email body support for both warnings and alerts
* Allow optional command line execution as the result of a threshold alert breach or restoral event.
* Allow threshold actions to be hooked by other plugins.  These hooks will be used by some of my non-public plugins for performing specific thold actions
* Add hooks for threshold and and template editing and saving
* Some faily significant code cleanup and code reduction
* In addition, I've found a number of smaller bugs that I will remediate via the GitHub issues list.

There are additional changes that I will be making subsequent to this commit, that include:

* Using a Threshold template to drive the creation of Graphs.  It's kind of backwards, but it works pretty well.
* Possible additional changes to the various hooks as I modify my gridalarms plugin to work with this version of thold.
* Any bug fixes that I discover from this major change.

Note, that you must be using Cacti 1.2 in order to leverage this new version of thold.  You are free to test it prior to it's release formal release of course.  I'm pretty sure all the core functionality is there as of right now.  However be warned, that I may for example pull or add a column or two, but you can rest assured that forcing another upgrade should fix things automatically.

Finally, let's not all get too excited.  I'm doing this work for my employer who has been kind enough to allow this work to happen.  None-the-less, I'm glad to be back, if only for a short time.

TW