Issue #435 - Notification Queue Implementation
This is the beginning of the notification queue.
It is not complete yet as the daemon and data collector must still be written.
Then the command line executor will have to be written as well.

Columns that have yet to be added include: environment, and process_id

It's likely that I will change the schema ever so slightly before I'm done as well.